fix bad cherry-pick
